編集系画面のリファクタリング
