+++ composables/use-customer-setting.ts
@@ -2,8 +2,13 @@ import type {
   PageSection,
   BasePageSection,
   PageSectionEdit,
-} from '@/types/page-setting'
+} from '@/types/customer-setting'
+import type { ColorTheme, LayoutTheme } from '@/types/customer'
 
+/**
+ * ホームレイアウト情報の取得処理
+ * @param customerId
+ */
 export const useHomeLayoutRead = (customerId: Ref<number | null>) => {
   const apiPath = '/layout/home'
   const domidPrefix = 'home-index'
@@ -52,6 +57,10 @@ export const useHomeLayoutRead = (customerId: Ref<number | null>) => {
   }
 }
 
+/**
+ * ホームレイアウト情報の更新・保存処理
+ * @param customerId
+ */
 export const useHomeLayoutWrite = (customerId: Ref<number | null>) => {
   const apiPath = '/layout/home'
   const { authorizationHeader } = useAuth()
@@ -93,6 +102,10 @@ export const useHomeLayoutWrite = (customerId: Ref<number | null>) => {
   }
 }
 
+/**
+ * ホームレイアウト情報変更用フォームデータ
+ * @param customerId
+ */
 export const useHomeLayoutEdit = (customerId: Ref<number | null>) => {
   const {
     homeSections,
@@ -158,7 +171,7 @@ export const useHomeLayoutEdit = (customerId: Ref<number | null>) => {
     nextKey()
     await replaceHomeLayout(editSections.value)
     await fetchHomeLayout()
-    addSnackber?.('ホームページレイアウトを変更しました。')
+    addSnackber?.('ホームページのレイアウトを変更しました。')
   }
 
   return {
@@ -166,5 +179,72 @@ export const useHomeLayoutEdit = (customerId: Ref<number | null>) => {
     editSections,
     loading,
     onUpdateSections,
+  }
+}
+
+/**
+ * 顧客情報のテーマ設定関連
+ */
+export const useThemeSettingsEdit = () => {
+  const { customer, setCustomer, updateCustomer, saving } = useCustomer()
+  const { addSnackber } = useSnackbars()
+
+  const editLayoutTheme = computed<LayoutTheme>({
+    get: () => customer.value?.layoutTheme ?? 'type1',
+    set: async (theme: LayoutTheme) => {
+      if (customer.value) {
+        const updateData = { ...customer.value, layoutTheme: theme }
+        setCustomer(updateData)
+        await updateCustomer(updateData)
+        addSnackber?.('レイアウトテーマを更新しました。')
+      }
+    },
+  })
+
+  const editColorTheme = computed<ColorTheme | undefined>({
+    get: () => customer.value?.colorTheme,
+    set: async (theme?: ColorTheme) => {
+      if (customer.value) {
+        const updateData = { ...customer.value, colorTheme: theme }
+        setCustomer(updateData)
+        await updateCustomer(updateData)
+        addSnackber?.('カラーテーマを更新しました。')
+      }
+    },
+  })
+
+  const layoutThemeOptions: {
+    type: LayoutTheme
+    label: string
+  }[] = [
+    {
+      type: 'type1',
+      label: '標準',
+    },
+    {
+      type: 'type2',
+      label: 'シャープ',
+    },
+  ]
+  const colorThemeOptions: {
+    type: ColorTheme
+    label: string
+  }[] = [
+    {
+      type: 'light',
+      label: 'ライト',
+    },
+    {
+      type: 'dark',
+      label: 'ダーク',
+    },
+  ]
+
+  return {
+    editLayoutTheme,
+    editColorTheme,
+    layoutThemeOptions,
+    colorThemeOptions,
+    saving,
   }
 }
